Update Content.js
Hey Mike, i was not able to make the add-on work, so i hacked together a version that polls for the video element itself. AFAIK, the video does not come in on load, and also, the class name u were originally using changed. Feel free to disregard, or improve on this

<3