Detect INXPO video via canplay event on capture.
It seems as if there was some kind of update in the INXPO playback software,
and we can't rely on the <video> element existing during the load event anymore.

This patch waits until the <video> element fires its canplay event, which means that
it has loaded enough information to actually play a few frames. At that point, the
speed control will attach.